# mcp-agents/contracts/mcp-agent-openapi.yaml
openapi: 3.1.0
info:
  title: NGKG MCP and Agent REST API
  version: 0.8.0
  description: Complete REST parity for NGKG semantic MCP tools, long-input compilation, managed agents, tenant tools, evidence-bound
    memory and distributed CPU qualification. Load this endpoint directly in Swagger UI.
servers:
- url: /
security:
- bearerAuth: []
...
paths:
...
  /v1/agent-inputs/{inputId}/parts/{ordinal}:
    put:
      tags:
      - Agent Inputs
      operationId: uploadAgentInputPart
      summary: Upload an idempotent input part
      parameters:
      - $ref: '#/components/parameters/inputId'
      - in: path
        name: ordinal
        required: true
        schema:
          type: integer
          minimum: 0
      responses:
        '200':
          description: Part receipt
      description: This operation implements the documented "Upload an idempotent input part" API capability. Use it when
        a client needs to ingest or inspect a large immutable agent input and its permanent constraint ledger. Inputs use
        the documented path, query and header parameters; success returns part receipt. Authentication, tenant authorization,
        resource ceilings and documented failure responses are enforced before a result is accepted.
  /v1/agent-inputs/{inputId}/finalize:
    post:
      tags:
      - Agent Inputs
      operationId: finalizeAgentInput
      summary: Freeze the exact input manifest and enqueue deterministic compilation
      parameters:
      - $ref: '#/components/parameters/inputId'
      responses:
        '200':
          description: Frozen manifest identity
      description: This operation implements the documented "Freeze the exact input manifest and enqueue deterministic compilation"
        API capability. Use it when a client needs to ingest or inspect a large immutable agent input and its permanent constraint
        ledger. Inputs use the documented path, query and header parameters; success returns frozen manifest identity. Authentication,
        tenant authorization, resource ceilings and documented failure responses are enforced before a result is accepted.
  /v1/agent-inputs/{inputId}:
    get:
      tags:
      - Agent Inputs
      operationId: getAgentInput
      parameters:
      - $ref: '#/components/parameters/inputId'
      responses:
        '200':
          description: Input status
      summary: get Agent Input
      description: This operation implements the documented "get Agent Input" API capability. Use it when a client needs to
        ingest or inspect a large immutable agent input and its permanent constraint ledger. Inputs use the documented path,
        query and header parameters; success returns input status. Authentication, tenant authorization, resource ceilings
        and documented failure responses are enforced before a result is accepted.
  /v1/agent-inputs/{inputId}/manifest:
    get:
      tags:
      - Agent Inputs
      operationId: getAgentInputManifest
      parameters:
      - $ref: '#/components/parameters/inputId'
      responses:
        '200':
          description: Checksum-bound manifest
      summary: get Agent Input Manifest
      description: This operation implements the documented "get Agent Input Manifest" API capability. Use it when a client
        needs to ingest or inspect a large immutable agent input and its permanent constraint ledger. Inputs use the documented
        path, query and header parameters; success returns checksum-bound manifest. Authentication, tenant authorization,
        resource ceilings and documented failure responses are enforced before a result is accepted.
  /v1/agent-inputs/{inputId}/requirements:
    get:
      tags:
      - Agent Inputs
      operationId: getAgentInputRequirements
      parameters:
      - $ref: '#/components/parameters/inputId'
      responses:
        '200':
          description: Constraint ledger
      summary: get Agent Input Requirements
      description: This operation implements the documented "get Agent Input Requirements" API capability. Use it when a client
        needs to ingest or inspect a large immutable agent input and its permanent constraint ledger. Inputs use the documented
        path, query and header parameters; success returns constraint ledger. Authentication, tenant authorization, resource
        ceilings and documented failure responses are enforced before a result is accepted.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    inputId:
      in: path
      name: inputId
      required: true
      schema:
        type: string
        format: uuid
...